Updated the wiki, finally finished gradle, moved resources, fixed TranslatedTest.
