+++ build.gradle
@@ -1,7 +1,3 @@
-/*
- * This file was generated by the Gradle 'init' task.
- */
-
 plugins {
     id 'java'
     id 'maven-publish'
@@ -44,6 +40,20 @@ publishing {
     }
 }
 
+configurations {
+    resolveImplementation
+    resolveImplementation.extendsFrom(implementation)
+}
+
+sourceSets {
+    test {
+        compileClasspath += configurations.resolveImplementation
+        compileClasspath += configurations.compileOnly
+        runtimeClasspath += configurations.resolveImplementation
+        runtimeClasspath += configurations.compileOnly
+    }
+}
+
 dependencies {
     compileOnly 'org.spigotmc:spigot-api:1.15.2-R0.1-SNAPSHOT'
     compileOnly 'co.aikar:taskchain-bukkit:3.7.2'
@@ -59,8 +69,6 @@ dependencies {
     testImplementation 'org.mockito:mockito-core:3.3.3'
     testImplementation 'org.powermock:powermock-module-junit4:2.0.7'
     testImplementation 'org.powermock:powermock-api-mockito2:2.0.7'
-    testImplementation 'org.spigotmc:spigot-api:1.15.2-R0.1-SNAPSHOT'
-    testImplementation 'com.google.code.findbugs:jsr305:3.0.2'
 
     compileOnly 'org.projectlombok:lombok:1.18.12'
     annotationProcessor 'org.projectlombok:lombok:1.18.12'
