+++ settings.gradle
@@ -2,4 +2,4 @@
  * This file was generated by the Gradle 'init' task.
  */
 
-rootProject.name = 'EnchantmentTokens-Api'
+rootProject.name = 'EnchantmentTokensAPI'
