+++ src/test/java/software/bigbade/utils/TranslateTest.java
@@ -1,0 +1,53 @@
+/*
+ * Addons for the Custom Enchantment API in Minecraft
+ * Copyright (C) 2020 BigBadE
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package software.bigbade.utils;
+
+import org.bukkit.Material;
+import org.bukkit.configuration.ConfigurationSection;
+import org.junit.Assert;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.powermock.core.classloader.annotations.PrepareForTest;
+import org.powermock.modules.junit4.PowerMockRunner;
+import software.bigbade.enchantmenttokens.localization.LocaleManager;
+import software.bigbade.enchantmenttokens.localization.TranslatedStringMessage;
+import software.bigbade.enchantmenttokens.utils.ItemUtils;
+
+import java.util.Collections;
+import java.util.Locale;
+
+import static org.powermock.api.mockito.PowerMockito.mock;
+import static org.powermock.api.mockito.PowerMockito.mockStatic;
+import static org.powermock.api.mockito.PowerMockito.when;
+
+@RunWith(PowerMockRunner.class)
+@PrepareForTest({ ItemUtils.class, LocaleManager.class, TranslatedStringMessage.class })
+public class TranslateTest {
+    @Test
+    public void testTranslate() {
+        mockStatic(ItemUtils.class);
+        when(ItemUtils.createItem(Material.BLACK_STAINED_GLASS_PANE, " ")).thenReturn(null);
+        
+        ConfigurationSection section = mock(ConfigurationSection.class);
+        when(section.get("country-language")).thenReturn("US");
+
+        LocaleManager.updateLocale(section, Collections.emptyList());
+        Assert.assertEquals("Test", new TranslatedStringMessage(Locale.US, "test").translate());
+    }
+}
